Syntrest refactor (#546)
Generic Get- and Post syntConsumers have been written. If an application requires a more complicated consumer, it should be written in the domain-package for the consumers, and extend the abstract SyntConsumer class. 
Integration tests for the get and post consumers have been written, and replaces the controller test for now. 
WebClient is now used for all external calls.