Add SLF4J logging to DoedsmeldingService #deploy-tps-messaging-service
Added a logging function to track sent statuses in DoedsmeldingService. With the help of @Slf4j annotation, all sent statuses for each environment are logged for better tracking and debugging. This allows a clearer understanding of the state of death notifications. Additionally, a refactor has been made in the SkdMeldingsheader class to simplify the construction of message header strings.