Refactor dead notification services in Endringsmelding service
Implemented updates in the Endringsmelding service to handle the sending and cancellation of dead notifications (`Doedsmelding`). This included updating logic in FoedselsmeldingRequestMapper to check for nulls, activating previously commented out methods in SkdEndringsmeldingerController, and changing return types from Flux to Mono in TpsMessagingConsumer and DoedsmeldingService. Also added a new DTO, DoedsmeldingResponseDTO, to handle responses.