Add status check and logging for address history service #deploy-tps-messaging-service
A filter has been added to DoedsmeldingService to only proceed with address data if the response is ok. In the tpsPersoner map entry condition check in AdressehistorikkService, it now checks for a non-null TpsPersonData instead of a non-null TpsSvar. A helper method 'isOk' is added in AdressehistorikkDTO to check the response status. Additionally, logging is introduced in TpsMessagingConsumer to log received address history responses. An application, 'testnav-adresse-service', is also added in the config file.