Add exception handling and logging to TpsMessagingConsumer #deploy-tps-messaging-service
Updated methods in the `TpsMessagingConsumer` class to include logs capturing the status of the cancellation and sending of the death report.  Additionally, transitions in the handling of death reports in `DoedsmeldingService` are amended to return-response based on existence of a person and enhanced error handling. Also added `IdentMiljoeDTO` to enrich `FoedselsmeldingService` for checking existence of a person.