Set responseTimeout for HttpClientRequest in Pensjonforvalter commands
Implemented a responseTimeout in the httpClientRequests for different commands in the Pensjonforvalter package. A constant REQUEST_DURATION of 30000 ms, was included in PensjonforvalterConsumer to be used as the timeout period. This enhancement will ensure that the application doesn't hang while waiting for responses if a server is slow or unresponsive.