Upgrade to reactive security settings in skattekort-service #deploy-skattekort-service
Upgraded the skattekort-service to use reactive security settings, as well as implementing OpenAPI config filters. Skattekortservice is now aligned with other services by using the reactive core library, the reactive security library, and the security core library. Additional updates include configuration alterations in the gradle file, and changes to the workflow in the .yml file.