Remove TPS Forvalteren proxy application files and update URLs (#3457)
This commit removes all TPS Forvalteren proxy related application files from the repository. This includes application resource configurations, test files, Java classes, and Docker-related files. This is most likely part of a refactoring process or due to the migration of the TPS Forvalteren proxy functionality to a different application or service.