Refktorert Kommuner2024
The logic for filtering "kommune" codes based on the year 2024 has been shifted to the KodeverkConsumer class from the Kodeverk controller class. A utility function filterKommuner2024 has also been updated to work with KodeverkBetydningerResponse type instead of KodeverkDTO and KodeverkAdjustedDTO. This change improves the structure and organization of the code.