Renamed 'bestKriterier' to 'bestilling' in multiple files #deploy-test-dolly-backend
The variable 'bestKriterier' was renamed to 'bestilling' in several Java and TypeScript files to improve code clarity. Additionally, the backend API path was updated in useMaler.tsx and Minside.cy.ts, as well as some logging added to MalPanel.tsx for testing. New fields were added to DollyRequest2MalBestillingMappingStrategy.java for greater functionality.