Refactor code and add new feature to create template from person
Deleted the unwanted console.log in MalPanel.tsx and refactored the mapping of 'maler' to use 'bestilling' instead of 'bestKriterier'. In DollyEndpoints.tsx, modified URLs to a different endpoint and added a new method 'opprettMalFraPerson'. Further, in PersonVisning.tsx, imported and implemented 'OpprettMalFraPerson' and made a minor refactor. Lastly, in DollyService.tsx, the method 'opprettMalFraPerson' was implemented.