Update "create template" modal with topic context #deploy-test-frontend
The commit adds a switch case to the MalModal.tsx file to determine a 'topic' based on the 'malType'. It changes modal title to include the selected topic when creating a new template. Also, buttons with the text "OPPRETT NY MAL" have been updated to just say "OPPRETT MAL" in the PersonVisning.tsx and BestillingDetaljer.tsx files.