Refactor DoedsmeldingService error handling #deploy-endringsmelding-service
Error handling in DoedsmeldingService has been refactored to improve clarity and accuracy. The check for "p" environment has been removed and the code has been streamlined to directly send a death cancellation message only if a person is marked as deceased. This helps to accurately reflect the status of identity handling in the requested environment