Remove "Opphold annet sted" address type in PDL functionalities #deploy-test-frontend
The "Opphold annet sted" address type was removed from person and environment data in the PDL services. This includes all references and UI components related to this address type. Additionally, the conditions for setting the 'master' field of different address types were adjusted according to the selected identity master or type.