Refactor code to handle NPIDs and implement NPID merge functionality
The commit introduces handling of NPIDs (Non-personal identifiers) and implements a method to merge NPIDs. A new DTO called MergeElement was created for this purpose. The methods we refactored include npidMerge and getOrdrer in PdlOrdreService, and createPerson in CreatePersonService. In addition, LocalDateTime.now() was replaced with now() in multiple files, and a new SwopIdentsService was also implemented.