Refactor initialDoedsfall export and update value presentation
Change export method for initialDoedsfall in initialValues.tsx to a function that accepts an optional master argument. Also extended metadata master value presentation in multiple components to include 'Doedsfall', 'PdlPersonInfo', 'Foedsel' and 'Persondetaljer'. In Personinformasjon.tsx, the add method for 'doedsdato' attribute was updated to make use of the new getInitialDoedsfall function.