Add gender attribute and update immigrant condition message
Added gender attribute to Personinformasjon.tsx and updated the condition message for identifying immigrants with `FNR` instead of `DNR` or `NPID`. Refactored the attribute addition function in gender to consider both `PDL` and `NPID`. Also, in KjoennForm, redefined the prop `kanVelgeMaster` instead of `identtype` to ensure flexibility in master selection.