Refactored person-related forms and updated validation
Significant refactor of the forms related to person data was carried out. This included changes to the validation rules of "fullmakt" form, with addition of a check for TestNorge identification in the fullmakt's motpart's personident field. Furthermore, streamlined initial form set-up in the "Personinformasjon" form was implemented to reduce redundancies. Changes were also made across multiple other components and accompanying tests to reflect the updates.