Refactor validation methods in SivilstandService and FullmaktService #deploy-test-pdl-forvalter
Updated several services and the associated tests. In SivilstandService, the validate method was refactored to include the person's data. In FullmaktService, updated the method to use StringUtils' isNotBlank function for readability. Added an extra parameter to the the validate function in the test files to match the updates in the SivilstandService.