Add functionality for saving motpartsPersonident in FullmaktService #deploy-test-pdl-forvalter
The FullmaktService.java file has been updated to include functionality for saving 'motpartsPersonident' if it doesn't exist in the personRepository. An AtomicReference to "DbPerson" has been added to hold the motpartPerson object, and LocalDateTime is used to set the last update time. This ensures that the 'motpartsPersonident' data is always available when required.