Corrected error in handling deletion against ident-pool.
A HendelseIdRequest class was introduced, and adjustments were made in PdlTestdataConsumer.java, PersonService.java, and HendelseIdService.java. These changes collectively improved the process of deleting identities from the pool, thereby correcting the previously existing error. A different strategy to generate and handle incidents in the HendelseIdService class was followed for optimization.