Refactor Base64 encoding in SokosPostCommand and SkattekortService classes #deploy-test-pdl-forvalter
In the SokosPostCommand class, removed the logging line related to Base64 decoding, which seemed unnecessary, and instead, Base64 encoding is now adapted to manage MIME types. For this purpose, StandardCharsets class was imported. Likewise minor changes were applied to SkattekortService class. The goal is to improve data encoding and transmission.