Update data types and add Skattekort validation utility #deploy-skattekort-service
Changed the datatype of 'inntektsar' from Integer to String in SokosRequest.java. A new Java class 'SkattekortValidator' was created to validate various aspects of a Skattekort, such as Arbeitstag, Arbeitgeber, and Trekktype. Furthermore, the skattekort validation method in SkattekortService.java was replaced with the newly created SkattekortValidator. In addition, isEmpty and isAmbiguous methods were added in SkattekortTilArbeidsgiverDTO.java for additional validation.