Remove SokosGetRequest class and refactor associated methods #deploy-skattekort-service
The SokosGetRequest class has been deleted and its usage in various methods has been directly replaced with the required parameters: 'ident' and 'inntektsaar'. This change streamlines the process of requesting a SkatteKort, removing the need to create a SokosGetRequest object first. It also improves the readability and manageability of the code.