Refactor SkattekortService to improve response handling #deploy-skattekort-service
The SkattekortService code was updated to improve how Skattekort responses are processed. Previously, a SokosResponse object was used, but this was removed and replaced with more direct handling using Base64 decoding and Flux mapping. In addition, the unmarshal method was tweaked to directly return a List of Arbeidsgiver objects. Changes made aim to streamline code and increase performance.