Refactor birth-related methods in multiple services #deploy-test-pdl-forvalter
The commit revolves around refactoring code in 'ForelderBarnRelasjonService', 'ForeldreansvarService' and 'MetadataTidspunkterService'. It replaces the usage of 'FoedselDTO' with 'FoedselsdatoDTO' and adds handling for an extra parameter 'Foedselsdato'. Exception checks are made more comprehensive in 'ForeldreansvarService'. These modifications improve the handling of birth-related information like birthdate and birthplace in the services.