Add Ansvarssubjekt field and improve handling in ForeldreansvarService #deploy-test-pdl-forvalter #deploy-test-dolly-backend
This update adds the new field ‘Ansvarssubjekt’ in the foreldreansvar section of the interface as well as the backend. Moreover, it introduces new conditions in the 'PdlEksisterendePerson' module to properly filter given the ‘Ansvarssubjekt’ field. This change improves the handling of Foreldreansvar data and makes it compatible with the updated requirements.