Feature/nytt endepunkt for navpersonidentifikator (#3530)
* Add NAVsPersonIdentifikator to PdlArtifact and PersonDTO

A new field, NAVsPersonIdentifikator, was added to PdlArtifact.java. Also created a new file, NAVsPersonIdentifikatorDTO.java, as part of this. The PersonDTO.java was updated to include a list of NAVsPersonIdentifikatorDTO. This is done to provide more precise identification capabilities.

* Implement support for NAVsPersonIdentifikator in services

Integrated the NAVsPersonIdentifikator into various services and mapping files. This includes updates in the data transformation services and additional logic to handle this new field. This identifier is added to enhance identification capabilities for specific user cases that rely on it.