Feature/sykemelding forbedre status tilbakemelding (#3681)
* Refactor exception handling and update sykemelding responses

Replaced custom exceptions with ResponseStatusException for better HTTP response handling. Updated service and controller methods to return SykemeldningResponseDTO for more informative responses. Added SykemeldningResponseDTO class in the DTO library.

* Refactor exception handling in SykemeldingConsumer

Replaced custom exceptions with ResponseStatusException to standardize error handling across the application. Modified relevant unit tests and command logic to align with this change.

* Add test and prod configurations for sykemelding API #deploy-test-sykemelding-api

Created a new `config-test.yml` for the development environment and a new `application-prod.yml` for the production environment. Updated the Dockerfile and workflow to support these configurations, including enabling TokenX for production.

* Rename config-test.yml to config.test.yml #deploy-test-sykemelding-api

This change standardizes the naming convention for configuration files. It ensures consistency across the project and simplifies future maintenance.

* Add SPRING_PROFILES_ACTIVE env variable to configs #deploy-test-sykemelding-api

This commit introduces the SPRING_PROFILES_ACTIVE environment variable to the test and production configuration files. It sets the value to 'dev' for the test environment and 'prod' for the production environment to ensure appropriate profile activation.

* Change profile annotation for LocalVaultConfig #deploy-test-sykemelding-api

Updated the @Profile annotation from "dev" to "local" in LocalVaultConfig class. This change ensures that the configuration is applied only when the "local" profile is active.

* Refactor profiles and update URLs for local and dev environments      #deploy-test-synt-sykemelding-api #deploy-test-sykemelding-api

Updated the configuration to differentiate between 'local' and 'dev' profiles, ensuring appropriate URL endpoints for each environment. This includes changes in the OAuth2 setup, Spring profiles, and Dockerfile adjustments. Added new configuration files to support "prod" and "local" environments.

* Add new application to proxy config
#deploy-pdl-proxy

Included the testnav-synt-sykemelding-api-dev application under the dev-gcp cluster in the proxy configuration file. This change ensures proper routing and access for the new development environment.

* Add test environment configs for synt-sykemelding-api-dev
#deploy-test-dolly-backend #deploy-arbeidsforhold-service

Added new application entries for 'testnav-synt-sykemelding-api-dev' in both arbeidsforhold-service and dolly-backend configuration files. This change ensures the new environment is recognized and integrated into the existing test setup.

* Add testnav-synt-sykemelding-api-dev to config file #deploy-organisasjon-service

Updated the configurational dependencies to include the new application testnav-synt-sykemelding-api-dev. This change is necessary for the development environment setup and ensures that all relevant services are listed for proper integration.

* Update logging config for dev and local profiles #deploy-test-sykemelding-api #deploy-test-synt-sykemelding-api

Expanded the "prod" profile to include "dev" for consistent logging and changed "dev" to "local" to better differentiate environments. This will improve clarity and maintainability of the logging configuration.

* Update application-local.yaml with sykemelding API dev config

Renamed 'testnav-synt-sykemelding-api' to 'testnav-synt-sykemelding-api-dev' and added 'testnav-pdl-forvalter-dev' name. This ensures proper configuration for the development environment.

* Refactor sykemelding classes and fix DTO typo #deploy-test-dolly-backend
#deploy-test-sykemelding-api #deploy-test-synt-sykemelding-api

Refactored the Sykemelding classes to use a new response DTO and improved logging for better traceability. Also corrected the typo in SykemeldingResponseDTO across the codebase, ensuring consistent naming conventions. Updated application URLs in the configuration files.

* Remove unused logging and add new API configuration. #deploy-test-dolly-backend

Eliminated unused logging of sykemelding response status for clarity and simplicity. Added configuration entries for testnav-sykemelding-api and testnav-synt-sykemelding-api to support new integrations.

* Deploy #deploy-arbeidsforhold-service #deploy-helsepersonell-service

* Update dependencies and API URLs for dev environment #deploy-test-sykemelding-api

Changed SpringDoc dependency from webflux to webmvc in build.gradle to match project requirements. Also corrected the URL for the testnav-synt-sykemelding-api in the local configuration file.

* Remove Avro schema dependencies
#deploy-test-sykemelding-api

Removed the inclusion of 'avro-schema' from both `settings.gradle` and `build.gradle`. The change involved eliminating related dependencies including Kafka clients and Avro serializers that are no longer required.

* Add testnav-sykemelding-api-dev to config.test.yml
#deploy-test-dolly-backend

This commit updates the configuration file to include the testnav-sykemelding-api-dev application. This ensures that the application is recognized and properly configured during testing environments.

* Add sykemeldingId for tracking submissions. #deploy-test-frontend

Incorporated 'sykemeldingId' to enable tracking of sykemelding submissions in the components `SyntSykemelding.tsx` and `DetaljertSykemelding.tsx`. This enhancement allows for improved data traceability and management. Additionally, code formatting was refined for uniformity and clarity.

* Reorder error message checks and fix sykemelding config
#deploy-test-dolly-backend

Reordered error message checks in ErrorStatusDecoder for more logical processing, checking for a 'message' before 'error'. Also corrected the configuration name for testnav-synt-sykemelding-api in application-local.yaml to ensure proper service references.

* Refactor variable name for clarity in KodeverkGetCommand
#deploy-test-dolly-backend

Renamed the variable "kodeverk" to "kodeverket" to improve code clarity and consistency with naming conventions. This change ensures better understanding and readability of the code, reducing potential confusion during future developments or reviews.