Bugfix/valideringsendepunkt (#3691)
* Add KontaktMedPasientDTO to SykemeldingDTO

Introduce the KontaktMedPasientDTO class to capture patient contact details such as contact date and reasons for no contact. This addition allows for enhanced tracking of patient interaction in the sykemelding domain.

* Add KontaktMedPasient mapping to sykemelding
#deploy-test-sykemelding-api

This update introduces a new class for handling KontaktMedPasient data, enriching the sykemelding model with contact information details. The changes ensure that KontaktMedPasient details are correctly included in the mapping strategy and are validated through updated test cases.

* Add test for handling null KontaktDato
#deploy-test-sykemelding-api

This commit introduces a new test case to verify the correct handling of null values for KontaktDato in the Sykemelding mapping logic. Additionally, it refactors the KontaktMedPasient constructor by removing the redundant start date parameter, streamlining object creation. These changes enhance the reliability and clarity of the codebase with regard to contact date management in health certificates.

* Handle null DTO in KontaktMedPasient constructor #deploy-test-sykemelding-api

Ensure that the KontaktMedPasient constructor checks for a null DTO before creating a new XMLHelseOpplysningerArbeidsuforhet.KontaktMedPasient instance. This change prevents potential null pointer exceptions and improves the robustness of the code.